restructure connection class
